+++ oci/database_exadata_iorm_config_test.go
@@ -1,0 +1,182 @@
+// Copyright (c) 2017, 2019, Oracle and/or its affiliates. All rights reserved.
+
+package provider
+
+import (
+	"fmt"
+	"testing"
+
+	"github.com/hashicorp/terraform/helper/resource"
+	"github.com/hashicorp/terraform/terraform"
+)
+
+var (
+	ExadataIormConfigResourceConfig = ExadataIormConfigResourceDependencies +
+		generateResourceFromRepresentationMap("oci_database_exadata_iorm_config", "test_exadata_iorm_config", Optional, Update, exadataIormConfigRepresentation)
+
+	exadataIormConfigSingularDataSourceRepresentation = map[string]interface{}{
+		"db_system_id": Representation{repType: Required, create: `${oci_database_db_system.t.id}`},
+	}
+
+	exadataIormConfigRepresentation = map[string]interface{}{
+		"db_system_id": Representation{repType: Required, create: `${oci_database_db_system.t.id}`},
+		"objective":    Representation{repType: Required, create: `AUTO`, update: `BALANCED`},
+		"db_plans":     RepresentationGroup{Required, dbPlanRepresentation},
+	}
+
+	dbPlanRepresentation = map[string]interface{}{
+		"db_name": Representation{repType: Required, create: `default`, update: `default`},
+		"share":   Representation{repType: Required, create: `1`, update: `2`},
+	}
+
+	ExadataIormConfigResourceDependencies = DefinedTagsDependencies + `
+
+	resource "oci_core_virtual_network" "t" {
+		compartment_id = "${var.compartment_id}"
+		cidr_block = "10.1.0.0/16"
+		display_name = "-tf-vcn"
+		dns_label = "tfvcn"
+	}
+	data "oci_identity_availability_domain" "ad" {
+		compartment_id 		= "${var.compartment_id}"
+		ad_number      		= 3
+	}
+	resource "oci_core_subnet" "exadata_subnet" {
+		availability_domain = "${data.oci_identity_availability_domain.ad.name}"
+		cidr_block          = "10.1.22.0/24"
+		display_name        = "ExadataSubnet"
+		compartment_id      = "${var.compartment_id}"
+		vcn_id              = "${oci_core_virtual_network.t.id}"
+		route_table_id      = "${oci_core_virtual_network.t.default_route_table_id}"
+		dhcp_options_id     = "${oci_core_virtual_network.t.default_dhcp_options_id}"
+		security_list_ids   = ["${oci_core_virtual_network.t.default_security_list_id}", "${oci_core_security_list.exadata_shapes_security_list.id}"]
+		dns_label           = "subnetexadata1"
+	}
+	resource "oci_core_subnet" "exadata_backup_subnet" {
+		availability_domain = "${data.oci_identity_availability_domain.ad.name}"
+		cidr_block          = "10.1.23.0/24"
+		display_name        = "ExadataBackupSubnet"
+		compartment_id      = "${var.compartment_id}"
+		vcn_id              = "${oci_core_virtual_network.t.id}"
+		route_table_id      = "${oci_core_virtual_network.t.default_route_table_id}"
+		dhcp_options_id     = "${oci_core_virtual_network.t.default_dhcp_options_id}"
+		security_list_ids   = ["${oci_core_virtual_network.t.default_security_list_id}"]
+		dns_label           = "subnetexadata2"
+	}
+
+	resource "oci_core_security_list" "exadata_shapes_security_list" {
+		compartment_id = "${var.compartment_id}"
+		vcn_id         = "${oci_core_virtual_network.t.id}"
+		display_name   = "ExadataSecurityList"
+
+		ingress_security_rules {
+			source    = "10.1.22.0/24"
+			protocol  = "6"
+		}
+
+		ingress_security_rules {
+			source    = "10.1.22.0/24"
+			protocol  = "1"
+		}
+
+		egress_security_rules {
+			destination = "10.1.22.0/24"
+			protocol    = "6"
+		}
+
+		egress_security_rules {
+			destination = "10.1.22.0/24"
+			protocol    = "1"
+		}
+	}
+
+	resource "oci_database_db_system" "t" {
+		availability_domain = "${data.oci_identity_availability_domain.ad.name}"
+		compartment_id = "${var.compartment_id}"
+		subnet_id = "${oci_core_subnet.exadata_subnet.id}"
+		backup_subnet_id = "${oci_core_subnet.exadata_backup_subnet.id}"
+		database_edition = "ENTERPRISE_EDITION_EXTREME_PERFORMANCE"
+		disk_redundancy = "NORMAL"
+		shape = "Exadata.Quarter1.84"
+		cpu_core_count = "22"
+		ssh_public_keys = ["ssh-rsa KKKLK3NzaC1yc2EAAAADAQABAAABAQC+UC9MFNA55NIVtKPIBCNw7++ACXhD0hx+Zyj25JfHykjz/QU3Q5FAU3DxDbVXyubgXfb/GJnrKRY8O4QDdvnZZRvQFFEOaApThAmCAM5MuFUIHdFvlqP+0W+ZQnmtDhwVe2NCfcmOrMuaPEgOKO3DOW6I/qOOdO691Xe2S9NgT9HhN0ZfFtEODVgvYulgXuCCXsJs+NUqcHAOxxFUmwkbPvYi0P0e2DT8JKeiOOC8VKUEgvVx+GKmqasm+Y6zHFW7vv3g2GstE1aRs3mttHRoC/JPM86PRyIxeWXEMzyG5wHqUu4XZpDbnWNxi6ugxnAGiL3CrIFdCgRNgHz5qS1l MustWin"]
+		domain = "${oci_core_subnet.exadata_subnet.dns_label}.${oci_core_virtual_network.t.dns_label}.oraclevcn.com"
+		hostname = "myOracleDB"
+		data_storage_size_in_gb = "256"
+		license_model = "LICENSE_INCLUDED"
+		node_count = "1"
+		time_zone = "US/Pacific"
+		db_home {
+			db_version = "12.1.0.2"
+			database {
+				"admin_password" = "BEstrO0ng_#11"
+				"db_name" = "aTFdb"
+			}
+		}
+	}
+	`
+)
+
+func TestDatabaseExadataIormConfigResource_basic(t *testing.T) {
+	provider := testAccProvider
+	config := testProviderConfig()
+
+	compartmentId := getEnvSettingWithBlankDefault("compartment_ocid")
+	compartmentIdVariableStr := fmt.Sprintf("variable \"compartment_id\" { default = \"%s\" }\n", compartmentId)
+
+	resourceName := "oci_database_exadata_iorm_config.test_exadata_iorm_config"
+
+	singularDatasourceName := "data.oci_database_exadata_iorm_config.test_exadata_iorm_config"
+
+	var resId, resId2 string
+
+	resource.Test(t, resource.TestCase{
+		PreCheck: func() { testAccPreCheck(t) },
+		Providers: map[string]terraform.ResourceProvider{
+			"oci": provider,
+		},
+		Steps: []resource.TestStep{
+			// verify create
+			{
+				Config: config + compartmentIdVariableStr + ExadataIormConfigResourceDependencies +
+					generateResourceFromRepresentationMap("oci_database_exadata_iorm_config", "test_exadata_iorm_config", Required, Create, exadataIormConfigRepresentation),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttrSet(resourceName, "db_system_id"),
+					resource.TestCheckResourceAttr(resourceName, "objective", "AUTO"),
+					resource.TestCheckResourceAttr(resourceName, "db_plans.#", "1"),
+					func(s *terraform.State) (err error) {
+						resId, err = fromInstanceState(s, resourceName, "id")
+						return err
+					},
+				),
+			},
+
+			// verify updates to updatable parameters
+			{
+				Config: config + compartmentIdVariableStr + ExadataIormConfigResourceDependencies +
+					generateResourceFromRepresentationMap("oci_database_exadata_iorm_config", "test_exadata_iorm_config", Required, Update, exadataIormConfigRepresentation),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttrSet(resourceName, "db_system_id"),
+					resource.TestCheckResourceAttr(resourceName, "objective", "BALANCED"),
+					resource.TestCheckResourceAttr(resourceName, "db_plans.#", "1"),
+					func(s *terraform.State) (err error) {
+						resId2, err = fromInstanceState(s, resourceName, "id")
+						if resId != resId2 {
+							return fmt.Errorf("Resource recreated when it was supposed to be updated.")
+						}
+						return err
+					},
+				),
+			},
+			// verify singular datasource
+			{
+				Config: config +
+					generateDataSourceFromRepresentationMap("oci_database_exadata_iorm_config", "test_exadata_iorm_config", Required, Create, exadataIormConfigSingularDataSourceRepresentation) +
+					compartmentIdVariableStr + ExadataIormConfigResourceConfig,
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttrSet(singularDatasourceName, "db_system_id"),
+				),
+			},
+		},
+	})
+}